pg_replslotdata - a tool for displaying replication slot information
The replication slots data is stored in binary format on the disk
under the pg_replslot/<<slot_name>> directory which isn't human
readable. If the server is crashed/down (for whatever reasons) and
unable to come up, currently there's no way for the
user/admin/developer to know what were all the replication slots
available at the time of server crash/down to figure out what's
the restart lsn, xid, two phase info or types of slots etc.

pg_replslotdata is a tool that interprets the replication slots
information and displays it onto the stdout even if the server is
crashed/down. The design of this tool is similar to other tools
available in the core today i.e. pg_controldata, pg_waldump.